feat: add a setup file
