+++ react-ts/src/__tests__/Gallery.test.tsx
@@ -3,29 +3,8 @@ import Gallery from '../components/gallery/Gallery';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { AppProvider } from '../components/context/AppState';
 import { Api } from '../util/enums';
-import pokemonListItems from './mockData/pokemonListItems.json';
 import pokemonListEmpty from './mockData/pokemonListEmpty.json';
-import pokemonDetails from './mockData/pokemonDetails.json';
-import pokemonItem from './mockData/pokemonItem.json';
 import { Mock } from 'vitest';
-
-global.fetch = vi.fn((url) => {
-  let responseData = {};
-
-  if (url.includes(`${Api.baseUrl}${Api.pokemonEndpoint}`)) {
-    responseData = pokemonDetails;
-  } else if (
-    new RegExp(`^${Api.baseUrl}${Api.speciesEndpoint}\\d+\\/$`).test(url)
-  ) {
-    responseData = pokemonItem;
-  } else if (new RegExp(`^${Api.baseUrl}${Api.speciesEndpoint}`).test(url)) {
-    responseData = pokemonListItems;
-  }
-
-  return Promise.resolve({
-    json: () => Promise.resolve(responseData),
-  });
-}) as Mock;
 
 describe('Gallery Component', () => {
   it('renders the specified number of cards', () => {
@@ -65,8 +44,4 @@ describe('Gallery Component', () => {
       return expect(screen.getByText('No cards available.')).toBeDefined();
     });
   });
-});
-
-afterEach(() => {
-  vi.restoreAllMocks();
 });
