+++ react-ts/vitest.config.ts
@@ -8,5 +8,6 @@ export default defineConfig({
     },
     environment: 'jsdom',
     globals: true,
+    setupFiles: ['./src/__tests__/setup.ts'],
   },
 });
